Make help regex a little more flexible (#22)
The regex that `harpoon help` uses was very strict about how it handled whitespace. In particular, it required exactly one whitespace character at the start of a line for it to recognize a command. If your IDE happens to indent with four spaces, tough luck.

This change allows multiple whitespace characters at the start of the line, and multiple whitespace characters before the comment (`## %% whatever`).